@@ -10,89 +10,74 @@
     "test:watch": "npm test -- --watch",
     "test:update": "npm test -- --u",
     "start": "babel-node server.js",
-    "build": "NODE_ENV=production webpack -p --progress --colors --config webpack.prod.config.js --bail",
+    "build": "NODE_ENV=production webpack --progress --config webpack.prod.config.js --bail",
     "lint": "eslint frontend",
     "coverage": "jest --coverage"
   },
   "dependencies": {
-    "@babel/cli": "~7.13.10",
-    "@babel/polyfill": "~7.8.3",
-    "@babel/preset-env": "~7.13.10",
-    "@babel/preset-react": "~7.12.13",
-    "@sentry/browser": "~6.2.2",
-    "@types/react": "~16.14.5",
-    "autoprefixer": "~9.8.6",
-    "axios": "~0.21.1",
-    "babel-loader": "~8.2.2",
-    "bootstrap": "~4.6.0",
-    "classnames": "~2.2.6",
-    "connected-react-router": "~6.9.1",
-    "cookie": "~0.4.0",
-    "css-loader": "~5.1.3",
-    "expose-loader": "~0.7.5",
-    "file-loader": "~6.2.0",
-    "history": "~4.10.1",
-    "imports-loader": "~0.8.0",
+    "@babel/cli": "~7.21.5",
+    "@reduxjs/toolkit": "~1.9.5",
+    "@sentry/browser": "~7.54.0",
+    "@types/react": "~18.2.8",
+    "axios": "~1.4.0",
+    "bootstrap": "~5.3.0",
+    "classnames": "~2.3.2",
+    "cookie": "~0.5.0",
     "lodash": "~4.17.21",
-    "marked": "~2.0.1",
-    "node-sass": "~4.14.1",
-    "postcss": "~8.2.8",
-    "postcss-loader": "~2.1.6",
-    "prop-types": "~15.7.2",
-    "react": "~16.14.0",
-    "react-bootstrap": "~1.5.2",
-    "react-dom": "~16.14.0",
-    "react-hot-loader": "~4.13.0",
-    "react-redux": "~7.2.3",
-    "react-router": "~5.2.0",
-    "redux": "~4.0.5",
-    "redux-devtools-extension": "~2.13.9",
-    "redux-thunk": "~2.3.0",
-    "resolve-url-loader": "~3.1.2",
-    "sass-loader": "~8.0.2",
-    "script-loader": "~0.7.2",
-    "style-loader": "~2.0.0",
-    "url-loader": "~4.1.1",
-    "webpack": "~4.46.0",
-    "webpack-bundle-tracker": "~1.0.0",
-    "webpack-dev-server": "~3.11.2",
-    "whatwg-fetch": "~3.6.2"
+    "marked": "~5.0.4",
+    "prop-types": "~15.8.1",
+    "react": "~18.2.0",
+    "react-bootstrap": "~2.7.4",
+    "react-dom": "~18.2.0",
+    "react-redux": "~8.0.7",
+    "react-refresh": "~0.14.0",
+    "react-router": "~6.12.0"
   },
   "devDependencies": {
     "@babel/core": "~7.13.10",
+    "@babel/eslint-parser": "~7.21.8",
+    "@babel/eslint-plugin": "~7.19.1",
     "@babel/node": "~7.13.10",
-    "@hot-loader/react-dom": "~16.14.0",
-    "ajv": "~6.12.6",
-    "babel-core": "~7.0.0-bridge.0",
-    "babel-eslint": "~10.1.0",
-    "babel-jest": "~26.6.3",
+    "@babel/preset-env": "~7.22.4",
+    "@babel/preset-react": "~7.22.3",
+    "@pmmmwh/react-refresh-webpack-plugin": "~0.5.10",
+    "@testing-library/jest-dom": "~5.16.5",
+    "@testing-library/react": "~14.0.0",
+    "@testing-library/user-event": "~14.4.3",
+    "ajv": "~8.12.0",
+    "autoprefixer": "~9.8.6",
+    "babel-jest": "~29.5.0",
+    "babel-loader": "~9.1.2",
     "circular-dependency-plugin": "~5.2.2",
-    "enzyme": "~3.11.0",
-    "enzyme-adapter-react-16": "~1.15.6",
-    "enzyme-to-json": "~3.6.1",
-    "eslint": "~6.8.0",
-    "eslint-config-prettier": "~6.15.0",
-    "eslint-config-vinta": "~1.4.5",
-    "eslint-import-resolver-webpack": "~0.13.0",
-    "eslint-plugin-babel": "~5.3.1",
-    "eslint-plugin-import": "~2.22.1",
-    "eslint-plugin-jest": "~22.21.0",
-    "eslint-plugin-jsx-a11y": "~6.4.1",
-    "eslint-plugin-prettier": "~3.3.1",
-    "eslint-plugin-promise": "~4.3.1",
-    "eslint-plugin-react": "~7.22.0",
-    "eslint-plugin-react-hooks": "~2.5.1",
-    "eslint-plugin-sonarjs": "~0.5.0",
-    "eslint-plugin-unicorn": "~15.0.1",
+    "css-loader": "~6.8.1",
+    "eslint": "~8.42.0",
+    "eslint-config-prettier": "~8.3.0",
+    "eslint-config-vinta": "~2.0.0",
+    "eslint-import-resolver-webpack": "~0.13.2",
+    "eslint-plugin-import": "~2.27.5",
+    "eslint-plugin-jest": "~26.0.0",
+    "eslint-plugin-jsx-a11y": "~6.7.1",
+    "eslint-plugin-prettier": "~4.2.1",
+    "eslint-plugin-promise": "~6.1.1",
+    "eslint-plugin-react": "~7.32.2",
+    "eslint-plugin-react-hooks": "~4.6.0",
+    "eslint-plugin-sonarjs": "~0.11.0",
+    "eslint-plugin-unicorn": "~40.1.0",
     "identity-obj-proxy": "~3.0.0",
-    "jest": "~26.6.3",
-    "jest-css-modules": "~2.1.0",
-    "mini-css-extract-plugin": "~0.12.0",
-    "prettier": "~1.19.1",
+    "jest": "~29.5.0",
+    "jest-environment-jsdom": "~29.5.0",
+    "mini-css-extract-plugin": "~2.7.6",
+    "postcss": "~8.4.24",
+    "postcss-loader": "~7.3.2",
+    "prettier": "~2.8.8",
     "redux-mock-store": "~1.5.4",
-    "webpack-cli": "~3.3.12"
-  },
-  "resolutions": {
-    "babel-core": "7.0.0-bridge.0"
+    "sass": "~1.62.1",
+    "sass-loader": "~13.3.1",
+    "style-loader": "~3.3.3",
+    "webpack": "~5.85.1",
+    "webpack-bundle-tracker": "~2.0.0",
+    "webpack-cli": "~5.1.3",
+    "webpack-dev-server": "~4.15.0",
+    "whatwg-fetch": "~3.6.2"
   }
 }
