@@ -1,6 +1,5 @@
 'use strict';
 
 module.exports = {
-  presets: ['@babel/preset-env', '@babel/preset-react'],
-  plugins: ['react-hot-loader/babel'],
+  presets: ['@babel/preset-env', '@babel/preset-react']
 };
