@@ -6,7 +6,11 @@ import './style.scss';
 const ColorDisplay = (props) => {
   const { color } = props;
 
-  return <span className={`color-${color}`}>{color}</span>;
+  return (
+    <span className={`color-${color}`} data-testid="color-display">
+      {color}
+    </span>
+  );
 };
 
 ColorDisplay.defaultProps = {
