@@ -1,6 +1,1 @@
-/* eslint-disable import/no-extraneous-dependencies */
-
-import enzyme from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-
-enzyme.configure({ adapter: new Adapter() });
+import '@testing-library/jest-dom';
